+++ src/org/loklak/LoklakServer.java
@@ -48,17 +48,17 @@ import org.loklak.api.server.HelloServlet;
 import org.loklak.api.server.PeersServlet;
 import org.loklak.api.server.ProxyServlet;
 import org.loklak.api.server.PushServlet;
-import org.loklak.api.server.GeoJsonPushServlet;
+import org.loklak.api.server.push.GeoJsonPushServlet;
 import org.loklak.api.server.SearchServlet;
 import org.loklak.api.server.SettingsServlet;
 import org.loklak.api.server.StatusServlet;
 import org.loklak.api.server.SuggestServlet;
 import org.loklak.api.server.AccountServlet;
 import org.loklak.api.server.ThreaddumpServlet;
-import org.loklak.api.server.FossasiaPushServlet;
-import org.loklak.api.server.OpenWifiMapPushServlet;
-import org.loklak.api.server.NodelistPushServlet;
-import org.loklak.api.server.FreifunkNodePushServlet;
+import org.loklak.api.server.push.FossasiaPushServlet;
+import org.loklak.api.server.push.OpenWifiMapPushServlet;
+import org.loklak.api.server.push.NodelistPushServlet;
+import org.loklak.api.server.push.FreifunkNodePushServlet;
 import org.loklak.data.DAO;
 import org.loklak.tools.Browser;
 import org.loklak.vis.server.MapServlet;
